pyproject.toml
[project]
name = "wagtail-ninja"
version = "0.1.16"
description = "Ninja API for Wagtail"
requires-python = ">=3.12"
readme="README.md"

dependencies = [
    "django-ninja>=1.4.3",
]

[dependency-groups]
dev = [
    "django-stubs>=5.2.2",
    "icecream>=2.1.5",
    "orjson>=3.11.2",
    "ruff>=0.12.1",
    "wagtail>=7.0.1",
    "wagtail-headless-preview>=0.8.0",
]
